Improve reading and place in fn
All the variables needed in reading
from the port were cluttering up the
namespace. Hence, we placed all that
into a function. Also, the reading
itself was improved, to be able to
read correctly, with the correct event
to wait for being set. Finally, we
wait for a character to be in the
buffer, then we print it, vs trying
to wait for 12 characters to print
them.