Use WaitCommEvent and ReadFile more efficiently
Instead of checking for a recieved char every
time, we might as well just attempt to read a
char until we get no values - which will help
reduce the number of calls to WaitCommEvent.
This change was directly inspired by this msdn
post:

https://docs.microsoft.com/en-us/previous-versions/
ff802693(v=msdn.10)?redirectedfrom=MSDN#conclusion

This reduction of calls to WaitCommEvent does in
fact create a difference in how well the data is
recieved by the program, in a small test in which
the serial device loops, printing many characters
and then waiting one second.